checks if files exist

+++ tools/CleanUp.sh
@@ -1,4 +1,18 @@
-rm -rf ./01_delimiters
-rm -rf ./02_encoding
 
-rm ./*.csv
+
+
+if [ -d "./01_delimiters" ] ; then
+	rm -rf ./01_delimiters
+fi
+
+if [ -d "./02_encoding" ] ; then
+	rm -rf ./02_encoding
+fi
+
+if [ -f "./*.csv" ] ; then
+	rm ./*.csv
+fi
+
+if [ -f "./*.csv" ] ; then
+	rm ./*.log
+fi
